Added confirmation step
Currently, the script interprets error cases for 'bundle show' as no 
gems needed, which has the irritating result of uninstalling all of 
your gems. For example when you have changed a gem version in your 
Gemfile, 'bundle show' gives the error "The bundle currently has 
whatever locked at 1.1.2., Try running `bundle update whatever`" and 
this causes gem-reset to uninstall all you gems.

This commit adds a confirmation step before the gems are removed as a 
workaround to this problem.